Initial Commit
This is just enough to make sure I have syntax highlighting in a simple
language. This follows the jetbrains tutorial on creating a language
plugin. I will continue through the tutorial to make sure I understand
the ins and outs of creating a full plugin, then adjust the grammar and
associated helpers to match the ccs grammar.